update readme instruction and relative test code

+++ src/com/esotericsoftware/reflectasm/util/NumberUtils.java
@@ -90,7 +90,7 @@ public abstract class NumberUtils {
             Object objects = isGetDistance ? null : Array.newInstance(toClass, Array.getLength(from));
             for (int i = 0; i < Array.getLength(from); i++) {
                 if (isGetDistance)
-                    distance = Math.min(distance, convertOrGetDistance(Array.get(from, i), toClass, true));
+                    distance = Math.min(distance, (int) convertOrGetDistance(Array.get(from, i), toClass, true));
                 else Array.set(objects, i, convertOrGetDistance(Array.get(from, i), toClass, false));
             }
             return (T) (isGetDistance ? distance : objects);
